requirements.txt
absl-py==0.15.0
six==1.16.0